Grammar Correction ("an" corrected to "a")
corrected headline text “an US” to “a US”
====
Explanation:
when using the articles “a” and “an,” usage choice is governed by
pronunciation not spelling.

If the word following the article begins ***with the sound of*** a
vowel, it should be preceded by “an.”

This is a common mistake, especially in print and text.

In this case, “US” is pronounced “You-Ess,” not “Ooo-Ess.”  The
“Y-sound” is not a vowel-sound, which means it is preceded by “a.”

Hope this helps!