Removed "Random Agent Spoofer"
„I'm not sure whether Random Agent Spoofer should be recommended any
more. It used to be a good add-on when it was able to spoof the time
zone, referrer IPs, and window size but it hasn't seen significant
development since the Firefox 45.0 release when the script injection
mechanism broke. Moreover, the add-on doesn't provide a method to
update the profiles as newer browser versions are released and this
means that if the author of the add-on doesn't update the add-on, the
users are stuck with the browser profiles that are shipped with the
add-on and which should be changed regularly as the browser is updated.“

https://www.reddit.com/r/privacytoolsIO/comments/5aoxki/suggestions_for_
several_changes_in/