StartMail and Ixquick.eu update
Email section: StartMail now accepts Bitcoin. Changed to "Accepted"

Search Engine section, worth mentioning:

Ixquick.eu is based in the Netherlands only. Not the US. (Ixquick
originated in the US, but was acquired by Surfboard Holding B.V, a
privately held Dutch corporation, in 2000. It is based in the
Netherlands, outside of US jurisdiction.)